hibák javítása

+++ src/main/java/com/benyovszki/szakdolgozat/service/impl/CategoryService.java
@@ -5,11 +5,13 @@ import com.benyovszki.szakdolgozat.exception.ErrorType;
 import com.benyovszki.szakdolgozat.exception.OperationException;
 import com.benyovszki.szakdolgozat.model.Category;
 import com.benyovszki.szakdolgozat.model.TransactionType;
+import com.benyovszki.szakdolgozat.model.user.Role;
 import com.benyovszki.szakdolgozat.repository.CategoryRepository;
 import com.benyovszki.szakdolgozat.util.EnumConverter;
 import dto.szakdolgozat.benyovszki.com.category.*;
 import lombok.AllArgsConstructor;
-import org.modelmapper.ModelMapper;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Service;
 import org.springframework.util.StringUtils;
 
@@ -19,6 +21,7 @@ import javax.persistence.criteria.*;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
+import java.util.Optional;
 
 @Service
 @AllArgsConstructor
@@ -26,18 +29,22 @@ public class CategoryService {
 
     private CategoryRepository categoryRepository;
     private EntityManager em;
+    private UserService userService;
 
     public Category findById(long id) {
-       return categoryRepository.findById(id)
-               .orElseThrow(() -> new OperationException(ErrorType.ENTITY_NOT_FOUND,
-                       String.format("No Category with this id: [ %s ]", id)));
+        return categoryRepository.findById(id)
+                .orElseThrow(() -> new OperationException(ErrorType.ENTITY_NOT_FOUND,
+                        String.format("No Category with this id: [ %s ]", id)));
     }
 
     public Category save(Category category) {
+        checkForAdminPrivilege(category);
         return categoryRepository.save(category);
     }
 
     public void delete(long id) {
+        Optional<Category> check = categoryRepository.findById(id);
+        checkForAdminPrivilege(check.orElse(null));
         categoryRepository.deleteById(id);
     }
 
@@ -47,7 +54,7 @@ public class CategoryService {
         CriteriaQuery<Category> cq = cb.createQuery(Category.class);
         Root<Category> categoryRoot = cq.from(Category.class);
         Order order = cb.asc(categoryRoot.get("name"));
-        cq.where(getPredicates(cb,categoryRoot, categoryQueryParams).toArray(new Predicate[0]));
+        cq.where(getPredicates(cb, categoryRoot, categoryQueryParams).toArray(new Predicate[0]));
         cq.orderBy(order);
 
         //count max result
@@ -67,17 +74,36 @@ public class CategoryService {
     private List<Predicate> getPredicates(CriteriaBuilder cb, Root<Category> rt, CategoryQueryParams categoryQueryParams) {
         List<Predicate> predicates = new ArrayList<>();
 
+        UserDetails userDetails =
+                (UserDetails)SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        Predicate globalPredicate = cb.isNull(rt.get("owner"));
+        Predicate username = cb.equal(rt.get("owner"), userService.getByUsername(userDetails.getUsername()));
+
+        predicates.add(cb.or(globalPredicate, username));
+
         if (Objects.isNull(categoryQueryParams)) {
             return new ArrayList<>();
         }
 
         if (StringUtils.hasText(categoryQueryParams.getName())) {
-            predicates.add(cb.like(rt.get("name"), "%" + categoryQueryParams.getName() + "%" ));
+            predicates.add(cb.like(rt.get("name"), "%" + categoryQueryParams.getName() + "%"));
         }
         if (Objects.nonNull(categoryQueryParams.getTransactionType())) {
             predicates.add(cb.equal(rt.get("transactionType"), EnumConverter.convert(categoryQueryParams.getTransactionType(), TransactionType.class)));
         }
 
         return predicates;
+    }
+
+    private void checkForAdminPrivilege(Category category) {
+
+        if(Objects.isNull(category)) {
+            return;
+        }
+
+        if (Objects.isNull(category.getOwner()) && !SecurityContextHolder.getContext().getAuthentication().getAuthorities().contains(Role.ROLE_ADMIN)) {
+            throw new OperationException(ErrorType.MODIFY_GLOBAL_TYPE, "Global type modification is only available for admin users");
+        }
+
     }
 }
